backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-multipart==0.0.9
pydantic==2.9.0
jinja2==3.1.4

# PII Detection
spacy==3.7.5
pytesseract==0.3.10
Pillow==10.4.0

# File parsing
PyMuPDF==1.24.0
python-docx==1.1.0

# ML / Model training
pandas==2.2.2
scikit-learn==1.5.1
numpy==1.26.4
